+++ src/sora_client_context.cpp
@@ -1,4 +1,4 @@
-#include "sora/sora_client_factory.h"
+#include "sora/sora_client_context.h"
 
 // WebRTC
 #include <api/create_peerconnection_factory.h>
@@ -25,22 +25,22 @@
 
 namespace sora {
 
-std::shared_ptr<SoraClientFactory> SoraClientFactory::Create(
-    const SoraClientFactoryConfig& config) {
-  std::shared_ptr<SoraClientFactory> f = std::make_shared<SoraClientFactory>();
+std::shared_ptr<SoraClientContext> SoraClientContext::Create(
+    const SoraClientContextConfig& config) {
+  std::shared_ptr<SoraClientContext> c = std::make_shared<SoraClientContext>();
 
-  f->config_ = config;
-  f->network_thread_ = rtc::Thread::CreateWithSocketServer();
-  f->network_thread_->Start();
-  f->worker_thread_ = rtc::Thread::Create();
-  f->worker_thread_->Start();
-  f->signaling_thread_ = rtc::Thread::Create();
-  f->signaling_thread_->Start();
+  c->config_ = config;
+  c->network_thread_ = rtc::Thread::CreateWithSocketServer();
+  c->network_thread_->Start();
+  c->worker_thread_ = rtc::Thread::Create();
+  c->worker_thread_->Start();
+  c->signaling_thread_ = rtc::Thread::Create();
+  c->signaling_thread_->Start();
 
   webrtc::PeerConnectionFactoryDependencies dependencies;
-  dependencies.network_thread = f->network_thread_.get();
-  dependencies.worker_thread = f->worker_thread_.get();
-  dependencies.signaling_thread = f->signaling_thread_.get();
+  dependencies.network_thread = c->network_thread_.get();
+  dependencies.worker_thread = c->worker_thread_.get();
+  dependencies.signaling_thread = c->signaling_thread_.get();
   dependencies.task_queue_factory = webrtc::CreateDefaultTaskQueueFactory();
   dependencies.call_factory = webrtc::CreateCallFactory();
   dependencies.event_log_factory =
@@ -52,16 +52,16 @@ std::shared_ptr<SoraClientFactory> SoraClientFactory::Create(
   // media_dependencies
   cricket::MediaEngineDependencies media_dependencies;
   media_dependencies.task_queue_factory = dependencies.task_queue_factory.get();
-  media_dependencies.adm = f->worker_thread_->BlockingCall([&] {
+  media_dependencies.adm = c->worker_thread_->BlockingCall([&] {
     sora::AudioDeviceModuleConfig config;
-    if (!f->config_.use_audio_device) {
+    if (!c->config_.use_audio_device) {
       config.audio_layer = webrtc::AudioDeviceModule::kDummyAudio;
     }
     config.task_queue_factory = dependencies.task_queue_factory.get();
     config.jni_env = sora::GetJNIEnv();
-    if (f->config_.get_android_application_context) {
+    if (c->config_.get_android_application_context) {
       config.application_context =
-          f->config_.get_android_application_context(config.jni_env);
+          c->config_.get_android_application_context(config.jni_env);
     }
     return sora::CreateAudioDeviceModule(config);
   });
@@ -74,7 +74,7 @@ std::shared_ptr<SoraClientFactory> SoraClientFactory::Create(
   auto cuda_context = sora::CudaContext::Create();
   {
     auto config =
-        f->config_.use_hardware_encoder
+        c->config_.use_hardware_encoder
             ? sora::GetDefaultVideoEncoderFactoryConfig(cuda_context, env)
             : sora::GetSoftwareOnlyVideoEncoderFactoryConfig();
     config.use_simulcast_adapter = true;
@@ -83,7 +83,7 @@ std::shared_ptr<SoraClientFactory> SoraClientFactory::Create(
   }
   {
     auto config =
-        f->config_.use_hardware_encoder
+        c->config_.use_hardware_encoder
             ? sora::GetDefaultVideoDecoderFactoryConfig(cuda_context, env)
             : sora::GetSoftwareOnlyVideoDecoderFactoryConfig();
     media_dependencies.video_decoder_factory =
@@ -94,21 +94,21 @@ std::shared_ptr<SoraClientFactory> SoraClientFactory::Create(
   media_dependencies.audio_processing =
       webrtc::AudioProcessingBuilder().Create();
 
-  if (f->config_.configure_media_dependencies) {
-    f->config_.configure_media_dependencies(media_dependencies);
+  if (c->config_.configure_media_dependencies) {
+    c->config_.configure_media_dependencies(media_dependencies);
   }
 
   dependencies.media_engine =
       cricket::CreateMediaEngine(std::move(media_dependencies));
 
-  if (f->config_.configure_dependencies) {
-    f->config_.configure_dependencies(dependencies);
+  if (c->config_.configure_dependencies) {
+    c->config_.configure_dependencies(dependencies);
   }
 
-  f->factory_ = sora::CreateModularPeerConnectionFactoryWithContext(
-      std::move(dependencies), f->connection_context_);
+  c->factory_ = sora::CreateModularPeerConnectionFactoryWithContext(
+      std::move(dependencies), c->connection_context_);
 
-  if (f->factory_ == nullptr) {
+  if (c->factory_ == nullptr) {
     RTC_LOG(LS_ERROR) << "Failed to create PeerConnectionFactory";
     return nullptr;
   }
@@ -117,9 +117,9 @@ std::shared_ptr<SoraClientFactory> SoraClientFactory::Create(
   factory_options.disable_encryption = false;
   factory_options.ssl_max_version = rtc::SSL_PROTOCOL_DTLS_12;
   factory_options.crypto_options.srtp.enable_gcm_crypto_suites = true;
-  f->factory_->SetOptions(factory_options);
+  c->factory_->SetOptions(factory_options);
 
-  return f;
+  return c;
 }
 
 }  // namespace sora